move flux ip adapter testing stuff to flux module of test
