+++ tests/pipelines/flux/test_pipeline_flux_inpaint.py
@@ -6,12 +6,8 @@ from transformers import AutoConfig, AutoTokenizer, CLIPTextConfig, CLIPTextMode
 from diffusers import AutoencoderKL, FlowMatchEulerDiscreteScheduler, FluxInpaintPipeline, FluxTransformer2DModel
 
 from ...testing_utils import floats_tensor, torch_device
-from ..testing_utils import (
-    BasePipelineTesterConfig,
-    FluxIPAdapterTesterMixin,
-    MemoryTesterMixin,
-    PipelineTesterMixin,
-)
+from ..testing_utils import BasePipelineTesterConfig, MemoryTesterMixin, PipelineTesterMixin
+from .testing_utils import FluxIPAdapterTesterMixin
 
 
 class FluxInpaintPipelineTesterConfig(BasePipelineTesterConfig):
