+++ tests/pipelines/testing_utils/__init__.py
@@ -8,7 +8,7 @@ from .cache import (
 )
 from .common import BasePipelineTesterConfig, PipelineTesterMixin, assert_mean_pixel_difference
 from .from_pipe import FromPipeTesterMixin
-from .ip_adapter import FluxIPAdapterTesterMixin, IPAdapterTesterMixin
+from .ip_adapter import IPAdapterTesterMixin
 from .lora import LoraMemoryTesterMixin, LoraTesterMixin, UNetLoraTesterMixin
 from .memory import (
     GroupOffloadTesterMixin,
@@ -29,7 +29,6 @@ __all__ = [
     "PipelineTesterMixin",
     "FromPipeTesterMixin",
     "IPAdapterTesterMixin",
-    "FluxIPAdapterTesterMixin",
     "LoraTesterMixin",
     "LoraMemoryTesterMixin",
     "UNetLoraTesterMixin",
